move rgb/hsl into runtime module
