+++ devs/run-tests/allcompile.ts
@@ -6,5 +6,6 @@ import "@devicescript/i2c"
 import "@devicescript/settings"
 import "@devicescript/drivers"
 import "@devicescript/gpio"
+import "@devicescript/runtime"
 
 console.log("all libraries compiled and didn't crash!")
