+++ package.json
@@ -17,13 +17,14 @@
         "devtools": "yarn build && node ./cli/devicescript devtools",
         "test:compiler": "node ./cli/devicescript ctool --test",
         "test:test": "cd packages/test && yarn test",
+        "test:runtime": "cd packages/runtime && yarn test",
         "test:cowsay": "cd packages/cowsay && yarn test",
         "test:observables": "cd packages/observables && yarn test",
         "test:cloud": "cd packages/cloud && yarn test",
         "test:settings": "cd packages/settings && yarn test",
         "test:i2c": "cd packages/i2c && yarn test",
         "test:snippets": "node ./cli/devicescript snippets",
-        "test": "yarn test:compiler && yarn test:test && yarn test:settings && yarn test:i2c && yarn test:observables && yarn test:cloud && yarn test:cowsay && yarn test:snippets",
+        "test": "yarn test:compiler && yarn test:test && yarn test:runtime && yarn test:settings && yarn test:i2c && yarn test:observables && yarn test:cloud && yarn test:cowsay && yarn test:snippets",
         "boards": "node ./cli/devicescript ctool --fetch-boards compiler/src/boards.json",
         "bump": "node scripts/bump.mjs",
         "hwdocs": "node scripts/genhwsupport.mjs",
