+++ packages/core/src/index.ts
@@ -8,4 +8,3 @@ import "./array"
 import "./events"
 import "./jacdac"
 import "./led"
-export { rgb, hsl } from "./led"
