+++ packages/runtime/src/main.ts
@@ -1,0 +1,8 @@
+import { describe, expect, test } from "@devicescript/test"
+import { rgb } from "."
+
+describe("rgb", () => {
+    test("0,0,0", () => {
+        expect(rgb(0, 0, 0)).toBe(0)
+    })
+})
