+++ website/docs/developer/packages/index.mdx
@@ -17,6 +17,7 @@ They do not need to be `npm install`ed.
 -   `@devicescript/cloud` - [cloud connectivity](/developer/cloud)
 -   `@devicescript/i2c` - interacting with [I2C peripherals](/developer/servers/i2c)
 -   `@devicescript/gpio` - working with [GPIO pins](/developer/servers/pins)
+-   `@devicescript/runtime` - runtime utilities
 -   `@devicescript/settings` - reading/writing [settings in flash memory](/developer/settings)
 -   `@devicescript/drivers` - [drivers for I2C peripherals](/developer/servers)
